Attempt to build against glib-rs 0.1.3
This is a newer version of glib-rs than the original crate was compiled
against. glib-rs 0.0.x fails to compile on Rust 1.22 (2017/11) or newer
and much older compilers are just broken in a modern environment.

There was a conflict between the libsecret-sys being loaded as private,
but the ffi module was exported as public. Defaulted it to private as
well.