FIXME: Revert this workaround for compilation errors
This comments out lines that breaks the lock/unlock calls for secret
items and collections. This will need to be reverted once the correct
code has been implemented to get this to compile on glib-rs.

With this workaround added, the crate compiled even with current stable
and nightly compilers using this older glib 0.3.1 crate.